big update with linking and css
i have linked up the main pages we need right now. vince is curretnly working on the login in page. once that is done i will link it to the other webpages. i have also changed the nav bar to fir the whole screen and make it look more neater. i need to add the iit banner so it gives it a more professional look